backend/pyproject.toml
[project]
name = "litigation-prep-backend"
version = "0.1.0"
description = "FastAPI backend for the litigation prep assistant"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.8.0",
  "sqlalchemy[asyncio]>=2.0.40",
  "uvicorn[standard]>=0.32.0",
  "openai>=1.0.0",
  "instructor>=1.7.0",
  "python-multipart>=0.0.9",
  "aiosqlite>=0.19.0",
  "pypdf>=5.0.0",
  "structlog>=24.0.0",
  "tenacity>=9.0.0",
  "PyJWT[crypto]>=2.8.0",
  "httpx>=0.27.0",
  "asyncpg>=0.31.0",
  "pinecone>=5.0.0",
  "langfuse>=4.5.0",
]

[project.scripts]
dev = "src.cli:dev"

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "anyio>=4.0.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
warn_unused_ignores = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
